Updated to accept file path by command line, moved code in App.cpp to main file
