--- include/App.h
@@ -1,11 +1,0 @@
-#ifndef APP_H
-#define APP_H
-
-/**
- * It's the App entry point. It calls the very
- * process of handling records, like reading them,
- * ediding and saving in hard disc.
- */
-void sequenceSetEntryPoint();
-
-#endif
